requirements.txt
Flask==2.0.1
Flask-CORS==3.0.10
python-dotenv==0.19.1
gunicorn==20.1.0
openai==0.27.0